test(plugin-kvm): make BGP announce test ASN constant a long
Companion to the 4-byte ASN fix: the shared CONFIGURED_ASN constant was
an int, which no longer autoboxes into the Long command parameter.

+++ plugins/hypervisors/kvm/src/test/java/com/cloud/hypervisor/kvm/resource/wrapper/LibvirtOvnBgpAnnounceCommandWrapperTest.java
@@ -44,7 +44,7 @@ import com.cloud.utils.script.Script;
 public class LibvirtOvnBgpAnnounceCommandWrapperTest {
 
     private static final String PUBLIC_IP = "217.179.89.42";
-    private static final int CONFIGURED_ASN = 24452;
+    private static final long CONFIGURED_ASN = 24452L;
 
     @Test
     public void announceWithExplicitAsnSkipsAutoDetectAndChainsConfigureRouterBgpNetwork() {
